move Snapshot class to snapshot_base.hpp
