+++ silkworm/db/snapshots/snapshot.hpp
@@ -16,12 +16,9 @@
 
 #pragma once
 
-#include <filesystem>
 #include <functional>
-#include <map>
 #include <memory>
 #include <optional>
-#include <string>
 #include <utility>
 
 #include <silkworm/core/common/base.hpp>
@@ -30,9 +27,8 @@
 #include <silkworm/core/types/block_body_for_storage.hpp>
 #include <silkworm/db/snapshots/path.hpp>
 #include <silkworm/db/snapshots/rec_split/rec_split_par.hpp>
-#include <silkworm/db/snapshots/seg/decompressor.hpp>
-#include <silkworm/infra/common/os.hpp>
 
+#include "snapshot_base.hpp"
 #include "snapshot_word_serializer.hpp"
 
 namespace silkworm::snapshots {
@@ -51,90 +47,6 @@ struct MappedTransactionsSnapshot {
     MemoryMappedRegion segment;
     MemoryMappedRegion tx_hash_index;
     MemoryMappedRegion tx_hash_2_block_index;
-};
-
-//! \brief Generic snapshot containing data points for a specific block interval [block_from, block_to).
-//! \warning The snapshot segment can also be externally managed. This means that the memory-mapping can happen
-//! outside of this class and a \code Snapshot instance can be created by specifying the \code MemoryMappedRegion
-//! segment containing the information about the memory region already mapped. This must be taken into account
-//! because we must avoid to memory-map it again.
-class Snapshot {
-  public:
-    class Iterator {
-      public:
-        using value_type = std::shared_ptr<SnapshotWordSerializer>;
-        using iterator_category = std::input_iterator_tag;
-        using difference_type = void;
-        using pointer = value_type*;
-        using reference = value_type&;
-
-        Iterator(
-            seg::Decompressor::Iterator it,
-            std::shared_ptr<SnapshotWordSerializer> serializer,
-            SnapshotPath path)
-            : it_(std::move(it)), serializer_(std::move(serializer)), path_(std::move(path)) {}
-
-        reference operator*() { return serializer_; }
-        pointer operator->() { return &serializer_; }
-
-        Iterator operator++(int) { return std::exchange(*this, ++Iterator{*this}); }
-        Iterator& operator++();
-
-        friend bool operator!=(const Iterator& lhs, const Iterator& rhs) = default;
-        friend bool operator==(const Iterator& lhs, const Iterator& rhs);
-
-      private:
-        seg::Decompressor::Iterator it_;
-        std::shared_ptr<SnapshotWordSerializer> serializer_;
-        SnapshotPath path_;
-    };
-
-    static inline const auto kPageSize{os::page_size()};
-
-    explicit Snapshot(SnapshotPath path, std::optional<MemoryMappedRegion> segment_region = std::nullopt);
-    virtual ~Snapshot() = default;
-
-    [[nodiscard]] SnapshotPath path() const { return path_; }
-    [[nodiscard]] std::filesystem::path fs_path() const { return path_.path(); }
-
-    [[nodiscard]] BlockNum block_from() const { return path_.block_from(); }
-    [[nodiscard]] BlockNum block_to() const { return path_.block_to(); }
-
-    [[nodiscard]] bool empty() const { return item_count() == 0; }
-    [[nodiscard]] std::size_t item_count() const { return decoder_.words_count(); }
-
-    [[nodiscard]] MemoryMappedRegion memory_file_region() const;
-
-    void reopen_segment();
-    virtual void reopen_index() = 0;
-
-    Iterator begin(std::shared_ptr<SnapshotWordSerializer> serializer) const;
-    Iterator end() const;
-
-    struct WordItem {
-        uint64_t position{0};
-        uint64_t offset{0};
-        Bytes value;
-
-        WordItem() {
-            value.reserve(kPageSize);
-        }
-    };
-    using WordItemFunc = std::function<bool(WordItem&)>;
-    bool for_each_item(const WordItemFunc& fn);
-
-    [[nodiscard]] std::optional<WordItem> next_item(uint64_t offset, ByteView prefix = {}) const;
-
-    void close();
-
-  protected:
-    void close_segment();
-    virtual void close_index() = 0;
-
-    //! The path of the segment file for this snapshot
-    SnapshotPath path_;
-
-    seg::Decompressor decoder_;
 };
 
 class HeaderSnapshot : public Snapshot {
